update level up functionality to work after combat and several ui updates, as well as fix the experience allocation based on spawnWeight

+++ src/game/actors/attributes/Attribute.java
@@ -27,7 +27,7 @@ public abstract class Attribute {
 
     @Override
     public String toString() {
-        return name + ": " + value;
+        return name + ": " + (int) value;
     }
 }
